Use RMagick to display image on graph
Gets list of colors in image with colormap, gets the color of each pixel in image and matches it to entry in colormap and creates a multi-dimesional array from the collected pixels based of rows and columns of image. Then writes the contents of the multi-dimesional array to csv file with a space delimeter. To be turned into one or more classes.